Add type casting to word count operations

+++ partials/_card-collection.php
@@ -79,7 +79,7 @@ if ( ! empty( $items ) ) {
       case 'fcn_chapter':
         if ( ! in_array( $item->ID, $processed_ids ) ) {
           $chapter_count += 1;
-          $word_count += get_post_meta( $item->ID, '_word_count', true );
+          $word_count += (int) get_post_meta( $item->ID, '_word_count', true );
           $processed_ids[] = $item->ID;
         }
         break;
@@ -118,7 +118,7 @@ if ( ! empty( $items ) ) {
 
             if ( ! in_array( $chapter->ID, $processed_ids ) ) {
               $chapter_count += 1;
-              $word_count += get_post_meta( $chapter->ID, '_word_count', true );
+              $word_count += (int) get_post_meta( $chapter->ID, '_word_count', true );
               $processed_ids[] = $chapter->ID;
             }
           }
